% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{corrdistapprox2}
\alias{corrdistapprox2}
\title{Meilleure approximation de la densité de rho chapeau, estimateur du coefficient de corrélation de Bravais-Pearson}
\usage{
corrdistapprox2(rho, rho_0, n)
}
\arguments{
\item{rho}{Valeur en laquelle la densité est évaluée}

\item{rho_0}{Valeur de référence pour le coefficient de corrélation de Bravais-Pearson}

\item{n}{Effectif de l'échantillon}
}
\value{
Numeric value
}
\description{
Meilleure approximation de la densité de rho chapeau, estimateur du coefficient de corrélation de Bravais-Pearson
}
\examples{
corrdistapprox2(.7,.8,30)

}
