% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{d_hotels_n}
\alias{d_hotels_n}
\title{Hotels}
\format{
Un objet \code{data.frame} avec huit variables et 39 observations :
\describe{
  \item{PAYS}{Factor w/ 5 levels, Pays où est situé l'hôtel}
  \item{ETOILE}{Integer vector, Catégorie de l'hôtel}
  \item{CONFORT}{Integer vector, Note de confort}
  \item{CHAMBRE}{Integer vector, Nombre de chambres}
  \item{CUISINE}{Integer vector, Note de la cuisine}
  \item{SPORT}{Integer vector, Note des instructures sprotives}
  \item{PLAGE}{Integer vector, Note de la plage}
  \item{PRIX}{Integer vector, Prix moyen pour une nuit}
}
}
\usage{
d_hotels_n
}
\description{
Jeu de données hotels (nom des hotels en rownames).
}
\keyword{datasets}
