% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{d_macdo}
\alias{d_macdo}
\title{MacDonald}
\format{
Un objet \code{data.frame} avec 24 variables et 260 observations :
\describe{
  \item{Category}{Factor w/ 9 levels, Type d'aliment}
  \item{Item}{Factor w/ 260 levels, Nom de l'aliment}
  \item{Serving.Size}{Factor w/ 107 levels, Taille}
  \item{Calories}{Integer vector, Calories}
  \item{Calories.from.Fat}{Integer vector, Calories dues à la graisse}
  \item{Total.Fat}{Numerical vector, Graisse totale}
  \item{Total.Fat....Daily.Value.}{Integer vector, \% de la quantité de graisse totale attendue par jour}
  \item{Saturated.Fat}{Numerical vector, Graisse saturée}
  \item{Saturated.Fat....Daily.Value.}{Integer vector, \% de la quantité de graisse saturée attendue par jour}
  \item{Trans.Fat}{Numerical vector, Trans.Fat}
  \item{Cholesterol}{Integer vector, Cholesterol}
  \item{Cholesterol....Daily.Value.}{Integer vector, \% de la quantité de cholesterol attendue par jour}
  \item{Sodium}{Integer vector, Sel}
  \item{Sodium....Daily.Value.}{Integer vector, \% de la quantité de sel attendue par jour}
  \item{Carbohydrates}{Integer vector, Carbohydrates}
  \item{Carbohydrates....Daily.Value.}{Integer vector, \% de la quantité de carbohydrates attendue par jour}
  \item{Dietary.Fiber}{Integer vector, Fibres}
  \item{Dietary.Fiber....Daily.Value.}{Integer vector, \% de la quantité de fibres attendue par jour}
  \item{Sugars}{Integer vector, Sucres}
  \item{Protein}{Integer vector, Protéines}
  \item{Vitamin.A....Daily.Value.}{Integer vector, \% de la quantité de vitamine A attendue par jour}
  \item{Vitamin.C....Daily.Value.}{Integer vector, \% de la quantité de vitamine C attendue par jour}
  \item{Calcium....Daily.Value.}{Integer vector, \% de la quantité de calcium attendue par jour}
  \item{Iron....Daily.Value.}{Integer vector, \% de la quantité de fer attendue par jour}
}
}
\source{
Kaggle.
}
\usage{
d_macdo
}
\description{
Jeu de données sur des aliments vendus au McDonald.
}
\details{
Ce jeu de données a été extrait de kaggle où il est possible de trouver une description complémentaire des variables.
}
\keyword{datasets}
