% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{d_pres2002}
\alias{d_pres2002}
\title{Elections presidentielles 2002 : premier tour}
\format{
Un objet \code{data.frame} avec 12 variables et 8 observations :
\describe{
  \item{Arlette.Laguiller}{Integer vector, Nombre de votes par région}
  \item{Dominique.Voynet}{Integer vector, Nombre de votes par région}
  \item{François.Bayrou}{Integer vector, Nombre de votes par région}
  \item{Frédéric.Nihous}{Integer vector, Nombre de votes par région}
  \item{Gérard.Schivardi}{Integer vector, Nombre de votes par région}
  \item{Jean.Marie.Le.Pen}{Integer vector, Nombre de votes par région}
  \item{José.Bové}{Integer vector, Nombre de votes par région}
  \item{Marie.George.Buffet}{Integer vector, Nombre de votes par région}
  \item{Nicolas.Sarkozy}{Integer vector, Nombre de votes par région}
  \item{Olivier.Besancenot}{Integer vector, Nombre de votes par région}
  \item{Philippe.de.Villiers}{Integer vector, Nombre de votes par région}
  \item{Ségolène.Royal}{Integer vector, Nombre de votes par région}
}
}
\usage{
d_pres2002
}
\description{
Jeu de données sur le premier tour des élections présidentielles de 2002.
Les noms des lignes indiquent les régions.
}
\keyword{datasets}
