% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{d_pres2007}
\alias{d_pres2007}
\title{Elections presidentielles 2007 : premier tour}
\format{Un objet \code{data.frame} avec 12 variables et 23 observations :
\describe{
  \item{Sarkozy}{Integer vector, Nombre de votes par région}
  \item{Bayrou}{Integer vector, Nombre de votes par région}
  \item{Royal}{Integer vector, Nombre de votes par région}
  \item{Le.Pen}{Integer vector, Nombre de votes par région}
  \item{Besanc.}{Integer vector, Nombre de votes par région}
  \item{VIlliers}{Integer vector, Nombre de votes par région}
  \item{Voynet}{Integer vector, Nombre de votes par région}
  \item{Laguiller}{Integer vector, Nombre de votes par région}
  \item{Bove}{Integer vector, Nombre de votes par région}
  \item{Buffet}{Integer vector, Nombre de votes par région}
  \item{Nihous}{Integer vector, Nombre de votes par région}
  \item{Schivardi}{Integer vector, Nombre de votes par région}
}}
\usage{
d_pres2007
}
\description{
Jeu de données sur le premier tour des élections présidentielles de 2002.
Les noms des lignes indiquent les régions.
}
\keyword{datasets}
