% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{d_vac}
\alias{d_vac}
\title{Lieux de vacances et occupation}
\format{
Un objet \code{data.frame} avec 8 lignes et 8 variables :
\describe{
  \item{\code{Hotel}}{Integer vector, Séjour à l'hotel}
  \item{\code{Location}}{Integer vector, Séjour en location}
  \item{\code{Res.Second}}{Integer vector, Séjour dans une résidence secondaire}
  \item{\code{Parents}}{Integer vector, Séjour chez les parents}
  \item{\code{Amis}}{Integer vector, Séjour chez des amis}
  \item{\code{Camping}}{Integer vector, Séjour au camping}
  \item{\code{Sej.org}}{Integer vector, Séjour organisé}
  \item{\code{Autres}}{Integer vector, Autres} 
}
}
\usage{
d_vac
}
\description{
Ce jeu de données croise le type de séjours de vacances avec l'occupation des individus. L'occupation des individus correspond au nom des lignes du jeu de données.
}
\keyword{datasets}
