% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{d_wow}
\alias{d_wow}
\title{Zones de World of Warcraft}
\format{Un objet \code{data.frame} avec 160 lignes et 12 variables :
\describe{
  \item{\code{Continent}}{Factor w/ 7 levels, Nom du continent}
  \item{\code{Area}}{Factor w/ 12 levels, Nom de la région}
  \item{\code{Zone}}{Factor w/ 80 levels, Nom de la zone}
  \item{\code{Subzone}}{Factor w/ 38 levels, Nom de la sous-zone}
  \item{\code{Type}}{Factor w/ 8 levels, Type de zone}
  \item{\code{Size}}{Integer vector, Taille}
  \item{\code{Controlled}}{Factor w/ 5 levels, Faction contrôlant la zone}
  \item{\code{Min_req_level}}{Integer vector, Niveau minimum requis pour pouvoir accéder à la zone}
  \item{\code{Min_rec_level}}{Integer vector, Niveau minimum recommandé pour s'aventurer dans la zone}
  \item{\code{Max_rec_level}}{Integer vector, Niveau maximum recommandé pour s'aventurer dans la zone}
  \item{\code{Min_bot_level}}{Integer vector, Niveau minimum pour pouvoir utiliser un bot dans la zone}
  \item{\code{Max_bot_level}}{Integer vector, Niveau maximum pour pouvoir utiliser un bot dans la zone} 
}}
\source{
Kaggle.
}
\usage{
d_wow
}
\description{
Ce jeu données est constitué des caractéristiques de zones du jeu World of Warcraft.
}
\details{
Ce jeu de données a été extrait de kaggle où il est possible de trouver une description complémentaire des variables.
}
\keyword{datasets}
