% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ecole2}
\alias{ecole2}
\title{Jeu de données école}
\format{Un objet \code{data.frame} avec 119 lignes et 2 variables :
\describe{
  \item{\code{Maths}}{Numerical vector, Note en mathématiques}
  \item{\code{Sport}}{Numerical vector, Note en sport} 
}}
\usage{
ecole2
}
\description{
Jeu de données école avec les résultats des tests de mathématiques et de sport. Ces données ont été simulées.
}
\keyword{datasets}
