% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ecole3}
\alias{ecole3}
\title{Jeu de données école (3 variables)}
\format{Un objet \code{data.frame} avec 119 lignes et 3 variables :
\describe{
  \item{\code{Maths}}{Numerical vector, Note en mathématiques}
  \item{\code{Sport}}{Numerical vector, Note en sport}
  \item{\code{Age}}{Numerical vector, Âge de l'élève} 
}}
\usage{
ecole3
}
\description{
Jeu de données école avec non seulement les résultats des tests de mathématiques et de sport mais aussi l'âge des élèves. Ces données ont été simulées.
}
\keyword{datasets}
