% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Gauss2F1}
\alias{Gauss2F1}
\title{Fonction hypergéométrique de Gauss (hypergeo)}
\usage{
Gauss2F1(a, b, c, x)
}
\arguments{
\item{a}{Paramètre pour Gauss2F1}

\item{b}{Paramètre pour Gauss2F1}

\item{c}{Paramètre pour Gauss2F1}

\item{x}{Argument principal, nombre réel ou complexe}
}
\value{
Numeric value
}
\description{
Fonction hypergéométrique de Gauss (hypergeo)
}
\examples{
Gauss2F1(1/2,1/2,(40-2)/2,1-.75^2)

}
