% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Gauss2F1gsl}
\alias{Gauss2F1gsl}
\title{Fonction hypergéométrique de Gauss (gsl)}
\usage{
Gauss2F1gsl(a, b, c, x)
}
\arguments{
\item{a}{Paramètre pour Gauss2F1}

\item{b}{Paramètre pour Gauss2F1}

\item{c}{Paramètre pour Gauss2F1}

\item{x}{Argument principal, nombre réel ou complexe}
}
\value{
Valeur numérique, l'évaluation de la fonction hypergéométrique de Gauss pour les paramètres \code{a}, \code{b}, \code{c} et \code{x}
}
\description{
Fonction hypergéométrique de Gauss (gsl)
}
\examples{
Gauss2F1gsl(1/2,1/2,(40-2)/2,1-.75^2)

}
