% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModStatR-package.R
\docType{package}
\name{ModStatR-package}
\alias{ModStatR-package}
\alias{ModStatR}
\title{ModStatR: Statistical Modelling in Action with R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Datasets and functions for the book "Modélisation statistique par la pratique avec R", F. Bertrand, E. Claeys and M. Maumy-Bertrand (2019, ISBN:9782100793525, Dunod, Paris). The first chapter of the book is dedicated to an introduction to the R statistical software. The second chapter deals with correlation analysis: Pearson, Spearman and Kendall simple, multiple and partial correlation coefficients. New wrapper functions for permutation tests or bootstrap of matrices of correlation are provided with the package. The third chapter is dedicated to data exploration with factorial analyses (PCA, CA, MCA, MDA) and clustering. The fourth chapter is dedicated to regression analysis: fitting and model diagnostics are detailed. The exercises focus on covariance analysis, logistic regression, Poisson regression, two-way analysis of variance for fixed or random factors. Various example datasets are shipped with the package: for instance on pokemon, world of warcraft, house tasks or food nutrition analyses.
}
\examples{
set.seed(314)
}
\references{
\emph{Modélisation statistique par la pratique avec R}, Frédéric Bertrand, Emmanuelle Claeys, Myriam Maumy-Bertrand, 2019, ISBN:9782100793525, Dunod, Paris, \url{https://www.dunod.com/sciences-techniques/modelisation-statistique-par-pratique-avec-r-cours-et-exercices-corriges}, \url{https://github.com/fbertran/ModStatR/} et \url{https://fbertran.github.io/ModStatR/}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fbertran.github.io/homepage/}
  \item \url{https://fbertran.github.io/ModStatR/}
  \item \url{https://github.com/fbertran/ModStatR/}
  \item Report bugs at \url{https://github.com/fbertran/ModStatR/issues/}
}

}
\author{
\strong{Maintainer}: Frederic Bertrand \email{frederic.bertrand@utt.fr} (\href{https://orcid.org/0000-0002-0837-8281}{ORCID})

Authors:
\itemize{
  \item Emmanuelle Claeys \email{emmanuelle.claeys@unistra.fr}
  \item Myriam Maumy-Bertrand \email{myriam.maumy@utt.fr} (\href{https://orcid.org/0000-0002-4615-1512}{ORCID})
}

}
\keyword{internal}
