% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModStatR-package.R
\docType{package}
\name{ModStatR}
\alias{ModStatR}
\alias{ModStatR-package}
\title{ModStatR}
\description{
Motivation: Package compagnon du livre Modélisation statistique par la pratique avec R. Il contient les codes des chapitres du livre ainsi que les solutions des exercices mais aussi d'autres compléments à découvrir.
}
\examples{
set.seed(314)
}
\references{
\emph{Modélisation statistique par la pratique avec R}, Frédéric Bertrand, Emmanuelle Claeys, Myriam Maumy-Bertrand, 2019, ISBN:9782100793525, Dunod, Paris, \url{https://www.dunod.com/sciences-techniques/modelisation-statistique-par-pratique-avec-r-cours-et-exercices-corriges}, \url{https://github.com/fbertran/ModStatR} et \url{https://fbertran.github.io/ModStatR}
}
