% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{parasites}
\alias{parasites}
\title{Parasites}
\format{
Un objet \code{data.frame} avec 12 lignes et 4 variables :
\describe{
  \item{\code{Total}}{Integer vector, Nombre total de parasites}
  \item{\code{N.morts}}{Integer vector, Nombre de parasites morts}
  \item{\code{Niveau.de.dose}}{Integer vector, Niveau de dose utilisé}
  \item{\code{Sexe}}{Factor w/ 2 levels, Sexe de l'hôte} 
}
}
\usage{
parasites
}
\description{
Jeu de données sur léthalité de la cypermethrine sur des parasites.
}
\keyword{datasets}
