% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{poke}
\alias{poke}
\title{Pokemons}
\format{
Un objet \code{data.frame} avec 800 lignes et 13 variables :
\describe{
  \item{\code{X.}}{Integer vector, PokeDex index number}
  \item{\code{Name}}{Factor w/ 800 levels, Name of the Pokemon}
  \item{\code{Type.1}}{Factor w/ 18 levels, Type of pokemon}
  \item{\code{Type.2}}{Factor w/ 18 levels, Other Type of Pokemon}
  \item{\code{Total}}{Integer vector, Sum of Attack, Sp. Atk, Defense, Sp. Def, Speed and HP}
  \item{\code{HP}}{Integer vector, Hit Points}
  \item{\code{Attack}}{Integer vector, Attack Strength}
  \item{\code{Defense}}{Integer vector, Defensive Strength}
  \item{\code{Sp..Atk}}{Integer vector, Special Attack Strength}
  \item{\code{Sp..Def}}{Integer vector, Special Defensive Strength}
  \item{\code{Speed}}{ Speed}
  \item{\code{Generation}}{Integer vector, Number of generation}
  \item{\code{Legendary}}{Factor w/ 2 levels, True if Legendary Pokemon False if not (more revision on mythical vs legendary needed)} 
}
}
\source{
Kaggle.
}
\usage{
poke
}
\description{
Jeu de données contenant les caractéristiques de pokemons. Les pokemons concernés apparaissent dans les jeux et non dans les cartes pokemon ou Pokemon Go..
}
\details{
Ce jeu de données a été extrait de kaggle où il est possible de trouver une description complémentaire des variables.
}
\keyword{datasets}
