% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ref.cor.mtest}
\alias{ref.cor.mtest}
\title{Test exact matriciel du corrélation de Bravais-Pearson avec une référence non nécessairement nulle}
\usage{
ref.cor.mtest(mat, matrho_0)
}
\arguments{
\item{mat}{Matrice des données}

\item{matrho_0}{Matrice des valeurs de référence pour chacun des coefficients de corrélation de Bravais-Pearson}
}
\value{
Liste comportant trois matrices : la matrice des p-valeurs, la matrice des coefficients de corrélations observés et la matrice des effectifs ayant servis au calcul de ces coefficients de corrélation
}
\description{
Test exact matriciel du corrélation de Bravais-Pearson avec une référence non nécessairement nulle
}
\examples{
data(Mesures5,package="BioStatR")
Mes5_red_lr = subset(Mesures5[,-5],subset=Mesures5$espece=="laurier rose")
ref.cor.mtest(Mes5_red_lr[,c("masse","taille","masse_sec")],0.7)

}
