% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ref.cor.test}
\alias{ref.cor.test}
\title{Test exact du coefficient de corrélation de Bravais-Pearson avec une référence non nécéssairement nulle}
\usage{
ref.cor.test(corobs, rho_0, n)
}
\arguments{
\item{corobs}{Valeur observée du coefficient de corrélation de Bravais-Pearson}

\item{rho_0}{Valeur de référence pour le coefficient de corrélation de Bravais-Pearson}

\item{n}{Effectif de l'échantillon}
}
\value{
Numeric value
}
\description{
Test exact du coefficient de corrélation de Bravais-Pearson avec une référence non nécéssairement nulle
}
\examples{
ref.cor.test(corobs=.7,rho_0=.8,n=30)

}
