% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{resistance}
\alias{resistance}
\title{Résistance d'un ciment}
\format{Un objet \code{data.frame} avec 36 lignes et 3 variables :
\describe{
  \item{\code{Melangeur}}{Factor w/ 3, COLUMN_DESCRIPTION}
  \item{\code{Casseur}}{Factor w/ 3, COLUMN_DESCRIPTION}
  \item{\code{Resistance}}{Integer vector, résistance en livres par pouces carrés} 
}}
\source{
Davies, O.L. et Goldsmith, P.L. (Eds.), \emph{Statistical Methods in Research and Production}, 4th edition, Oliver and Boyd, Edinburgh, 1972.
}
\usage{
resistance
}
\description{
Jeu de données contenant les résultats d'une expérience évaluant la résistance d'un ciment.
}
\details{
Davies et Goldsmith ont récolté les données d’une expérience dont le but était d’étudier les diﬀérentes sources de variabilité possibles de la résistance d’un ciment fabriqué à Portland. 
L’expérience s’est déroulée ainsi : plusieurs petits prélèvements d’un même type de ciment ont été mélangés à de l’eau et travaillés par trois personnes diﬀérentes, les « mélangeurs ». On a alors formé douze cubes à l’aide de chacune des préparations des « mélangeurs ». 
Puis on a donné ces 36 cubes à trois personnes chargées d’évaluer leur résistance, les « casseurs ». 
La répartition des 36 cubes entre ces « casseurs » a été faite de telle sorte que chaque « casseur » reçoive quatre cubes provenant de chacune des préparations des « mélangeurs » soit douze cubes au total.
Tous les tests de résistance ont été faits sur la même machine. L’objectif principal de cette expérience était d’étudier et de quantiﬁer l’importance de la variabilité dans les tests de résistance qui pouvait provenir des diﬀérences individuelles entre les « mélangeurs » et les « casseurs ». 
Les données ci-dessous, exprimées dans les unités d’origine c’est-à-dire .
}
\keyword{datasets}
