% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rho.mult}
\alias{rho.mult}
\title{Matrice de corrélation de Bravais-Pearson, bootstrap ou permutation}
\usage{
rho.mult(mat, indices)
}
\arguments{
\item{mat}{Matrice des données}

\item{indices}{Vecteur d'indices dont la longueur est égale au nombre de lignes de la matrice}
}
\value{
Matrice des corrélations de Bravais-Pearson des données permutées
}
\description{
Matrice de corrélation de Bravais-Pearson, bootstrap ou permutation
}
\examples{
data(Mesures5,package="BioStatR")
Mes5_red_gv = subset(Mesures5[,-5],subset=Mesures5$espece=="glycine violette")
set.seed(1133)
rho.mult(Mes5_red_gv[,c("masse","taille","masse_sec")],sample(nrow(Mes5_red_gv)))

}
