% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rho}
\alias{rho}
\title{Corrélation de Bravais-Pearson pour bootstrap ou permutation}
\usage{
rho(x, y, indices)
}
\arguments{
\item{x}{un vecteur numérique}

\item{y}{un vecteur numérique}

\item{indices}{un vecteur d'indices de même longueur que x et y}
}
\value{
Valeur numérique, le coefficient de corrélation de Bravais-Pearson calculé pour la permutation des vecteurs \code{x} et \code{y} spécifiée par le vecteur \code{indices}
}
\description{
Corrélation de Bravais-Pearson pour bootstrap ou permutation
}
\examples{
set.seed(1133)
rho(rnorm(30),rnorm(30),sample(30))

}
