% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SidaChat}
\alias{SidaChat}
\title{Sida du chat}
\format{
Un objet \code{data.frame} avec 33 lignes et 3 variables :
\describe{
  \item{\code{Sexe}}{Factor w/ 2 levels, Sexe de l'animal}
  \item{\code{Jours}}{Integer vector, Nombre de jours après l'inoculation}
  \item{\code{LnT4}}{Numerical vector, Logarithme népérien du taux de leucocytes T4} 
}
}
\source{
\emph{Modèle linéaire : Comparaison de groupes et régression} de B. Prum aux Éditions de l’INSERM, 1996.
}
\usage{
SidaChat
}
\description{
Taux de leucocytes
}
\details{
Le taux de leucocytes T4 chez le chat a été mesuré plusieurs jours (valeur de la variable Jours) après avoir inoculé à l’animal le virus FeLV, analogue au HIV. 
Nous appellons LnT4 le logarithme népérien de ce taux de leucocytes T4.
}
\keyword{datasets}
