% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{vitamines}
\alias{vitamines}
\title{Vitamines}
\format{Un objet \code{data.frame} avec 32 lignes et 3 variables :
\describe{
  \item{\code{Calorie}}{Integer vector, COLUMN_DESCRIPTION}
  \item{\code{Vitamine}}{Integer vector, COLUMN_DESCRIPTION}
  \item{\code{Poids}}{Integer vector, COLUMN_DESCRIPTION} 
}}
\source{
D’après B. Falissard. \emph{Comprendre et utiliser les statistiques dans les sciences de la vie}. Masson, 2005.
}
\usage{
vitamines
}
\description{
Influence de diﬀérents régimes alimentaires sur des rats de laboratoire.
}
\details{
Le gain de poids des rats est désigné par la variable Poids, exprimée en grammes, les deux facteurs sont les variables Calorie et Vitamine. 
La variable Calorie vaut 1 si les rats n’ont pas suivi un régime hypercalorique et 2 s’ils ont suivi un tel régime hypercalorique. 
La variable Vitamine vaut 1 si les rats n’ont pas reçu de compléments vitaminés et 2 s’ils ont reçu de tels compléments.
}
\keyword{datasets}
