% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{corrdist}
\alias{corrdist}
\title{Densité de rho chapeau, estimateur du coefficient de corrélation de Bravais-Pearson}
\usage{
corrdist(rho, rho_0, n)
}
\arguments{
\item{rho}{Valeur en laquelle la densité est évaluée}

\item{rho_0}{Valeur de référence pour le coefficient de corrélation de Bravais-Pearson}

\item{n}{Effectif de l'échantillon}
}
\value{
Numeric value
}
\description{
Densité de rho chapeau, estimateur du coefficient de corrélation de Bravais-Pearson
}
\examples{
corrdist(.7,.8,30)

}
